Allow enabling/disabling of LitterHopper
API refers to disabling as "removing" so the function signatures reflect this language.

Unfortunately there is an async delay between sending the mutation and seeing the LitterHopper status updated to reflect it in the regular robot query. More than 1s, but less than 5s. Because the mutation doesn't expose hopper-related robot properties, we must refresh with this delay.